💄[Design] 지도 핀 선택 확인 버튼 추가 및 장소 선택 UX 개선
- 지도 모드 하단 toolbar에 "이 위치로 선택" 확인 버튼 추가
- 지도 모드 내비게이션 서브타이틀 "핀을 눌러 위치를 선택하세요" 안내 추가
- 기존 선택 장소가 있으면 지도 진입 시 해당 위치를 초기 핀으로 표시
- InlineMapPlacePicker/InlineMapPickerState를 별도 파일로 분리
- SearchContent/MapMarkerIcon을 SearchPlaceResultRow.swift로 분리
- 시트 닫기만으로 자동 적용되던 방식 → 확인 버튼 확정 방식으로 변경
- 불필요한 테스트 파일 정리 (CommunityPostResponseDTOTests, CommunityPostViewModelTests, NoticeAuthorFallbackTests)
