+++ AppProduct/AppProductTests/HomeTest/ScheduleDDayFormattingTests.swift
@@ -2,7 +2,7 @@
 //  ScheduleDDayFormattingTests.swift
 //  AppProductTests
 //
-//  Created by Codex on 3/12/26.
+//  Created by euijjang97 on 3/12/26.
 //
 
 @testable import AppProduct
